Refactored/moved classes
